Add functions to extract.py
